requirements.txt
PyPDF2
autogen
streamlit
python-dotenv